// src/models/client-address.model.ts
export interface ClientAddressModel {
    address: string;
    addressName: string;
    city: string;
    client_id: number;
    comment: string;
    id: number;
    isDefault: boolean;
    latitude: number;
    longitude: number;
    name: string;
    phoneNumber: string;
    surname: string;
    zipCode: string;
}